Bump ruff to 0.16.0 and resolve its broadened default rule set
ruff 0.16.0 expands the default lint set, which this repo's extend-select = [I, B, UP] inherits. Bump both pins (CI and the pre-commit hook) to 0.16.0 and clear the resulting findings: mechanical rewrites for the safe rules (C408/C401/RUF015/RUF046/PLR1730/SIM102/SIM103/SIM114), removal of the noqa directives 0.16.0 reports as unused, and justified per-line noqa for the deliberate idioms (PLR0124 cftime-NaT test, DTZ007 date-only parse, BLE001 warn-and-continue catches, TRY004 malformed-frontmatter ValueError contract). Rule selection is unchanged.

+++ skills/coarsen/scripts/coarsen.py
@@ -44,8 +44,8 @@ def _target_axis(coord_vals, resolution: float, offset: float):
     # Tolerance on (vmin, vmax) - offset / resolution to keep boundary points
     # that are on-grid up to floating-point noise.
     eps = 1e-9 * max(1.0, abs(vmin), abs(vmax)) / resolution
-    k_min = int(math.ceil((vmin - offset) / resolution - eps))
-    k_max = int(math.floor((vmax - offset) / resolution + eps))
+    k_min = math.ceil((vmin - offset) / resolution - eps)
+    k_max = math.floor((vmax - offset) / resolution + eps)
     if k_max < k_min:
         raise ValueError(
             f"No grid points at offset={offset}, resolution={resolution} "
